Use mkdtemp() rather than mktemp()
We currently create the build directory by using mktemp() to find
a unqiue filename in /tmp and then use that filename as the name
of the directory to create in /var/tmp. Just weird, frankly.

That's just weird, frankly. We should use mkdtemp() instead.

Also, unconditionally delete the build dir.