From Elias Hunt <hunt AT m2s DOT com>
Here is a patch based on my conversation with Mark on supporting more
than 2 nameservers. I just changed it to effectively support an
unlimited number of nameservers. However I believe the linux network
stack may have a maximum it will use before failing a connection
attempt, I'm just not clear on what that number would be so went in
favor of infinite.