squash osmin
The attached patch implements the solution recently discussed, to avoid
wasting 1.2M of ram during LiveOS sessions, during the non-install times
when the 1.2M of ram for the minimized filesystem overlay is not
actually needed.

This implementation involves putting osmin in its own squashfs, and
copying that 12kb image to ram, as opposed to an alternative of putting
osmin in the same squashfs as the main system image.  I chose to
implement the seperate version, because it guarantees that the install
will remain one long ~2G seekless cdrom read.