Fix --base-on
I'm working with the latest version of livecd-creator available in rawhide,
which is livecd-tools-012-1.fc8.  I noticed a bug when using the --base-on
option; basically, it is testing for the existence of the squashfs.img, but that
path can't possibly exist until the ISO has been loopback mounted.  The attached
patch moves the mounting of the ISO before the check to see if the squashfs
filesystem exists.  With this patch, I am able to successfully use the --base-on
option.

Signed-off-by: Chris Lalancette <clalance@redhat.com>