Allow repository URLs to be overridden
When building images, you may want to build from a local
yum repository to speed things up. However, currently
you have to modify the kickstart file to change the
URL. If the file is under revision contol, that's a bit
of a pain.

This patch allows installPackages() to take a dict of
repository URLs allowing a --override-repo type option
to be implemented.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>