Fix baseOnIso() stuff
Currently, it looks like baseOnIso() just copies the image
from the iso and then doesn't mount it.

First, we need to chain up to LoopImageCreator's
_mountInstallRoot() once we've copied the image.

Second, we need to make sure that SparseExt3LoopbackMount
doesn't try and create the image if it already exists. The
old code used to use LoopbackMount directly for this, but
then you loose the info about what filesystem it should be.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>