Introduce a name attribute
Since an fslabel needs to be less than 32 chars, introduce
a name property with no length restrictions.

Use the name property anywhere there is no need for a limited
length label.

Also, if the fslabel is not specified, default to just
truncating the name to 32 chars.

Finally, add a build_name() utility function which uses the
logic from livecd-creator and expose the maximum length for
fslabel via the FSLABEL_MAXLEN module attribute.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>