Add support for pxeboot
Myself and Chris Lalancette looked into this in some detail and came up
with two different ways to do this.  This is the simpler of the two
methods and results in the least invasive changes to livecd-creator --
just a single line patch to mayflower and a 100-or-so-line extra script
called 'livecd-iso-to-pxeboot' which works analogously to
'livecd-iso-to-disk'.

The way it works is to bundle the complete ISO image inside the initrd.
The kernel and (bloated) initrd are downloaded using PXE in the normal
way, and the init script finds and loopback-mounts the ISO image and
booting continues as normal.