LiveCD - add test for /sbin/lokkit if it is needed for selinux config
This patch adds a new function ayum.installHasFile() which tells if the install
image is going to contain a given file.  We then use this new function to make
sure lokkit is going to be present so that we will be able to disable selinux
inside the image if the kickstart was configured that way.  If we cannot
accomidate the kickstart setting we error the build.

Signed-off-by: Eric Paris <eparis@redhat.com>