Re-factor imgcreate/fs.py module
Most of the change here involves re-factoring the imgcreate/fs.py module to
separate the difference between a disk and a mount.

So this patch separates the roles. There are now classes which deal with
accessing / creating disks and classes which deal with mounting things.

Also fixed comments to use the standard pydoc """ format.
Based on a patch by Daniel Berrange <berrange@redhat.com>