And remove current configs
Live image configs for the distribution are now kept in the spin-kickstarts
git repository on fedorahosted.  This is because they really shouldn't be
incredibly tied to the tools, other than taking advantage as new features
are added.  It also makes it easier to delegate out commit access for the
configs.

Leaving the minimal config here so that there is something to have for
testing purposes