Temporarily patch up ramdisk handling for XO via olpc.fth (from Mitch Bradley)
The XO firmware is loading the initrd at the wrong address which is what
was breaking the use of standard Fedora kernels.  This does runtime
patching of ofw from the boot file until there's a new firmware
revision that we can require.

OFW code is from Mitch Bradley <wmb AT laptop DOT org>