Fix a bug with getting groups with options (@foo --optional).
New yum api (in 3.2.x) allows us to pass into selectGroups what set we
want from the group.  The default is mandatory and default.
Use that unless we have options from kickstart for either
optional packages (IE everything in the group) or just the
default packages of the group.

This also allows us to just get the right packages the first time
rather than getting packages and then pruning/adding as necessary.
Less trips through the yum stack, which is a good thing.