Using both EFI and MBR at once on Live USB images and other misc. fixes
Good news - syslinux's gptmbr.bin is working fine. I was using the partition
type UUID, not its unique UUID. With this correction, things work fine when
booting with BIOS or EFI!

I've attached a patch which re-adds the check for gptmbr.bin and if it's
found, then the script will now automatically grab the partition's unique
UUID and write a compatible MBR. It should work on all drives with an LBA
size of 512 bytes.

The patch also includes a minor fix: If --format is specified and the user
has accepted the first warning by hitting <enter>, then it will
automatically unmount all partitions on the device so that the parted
commands don't fail.