Tweak kernel/initrd paths for the multi-image case
When you load the extra syslinux.cfg, we need to be sure we reference
the kernel and initrd with full paths so the right ones are loaded.

Go ahead and set that up.

For those who want to try, add a section like the following to your
top-level syslinux/syslinux.cfg
  label otherimage
    menu label Other Image
    kernel vesamenu.c32
    APPEND /livedir/syslinux/syslinux.cfg