Only the UUID of the EFI System Partition is important, not its name
EFI systems should boot fine regardless of the name of the "EFI System
Partition" as it's defined by its UUID, not its label, which the boot
flag check already takes care of.

I verfied this on a MacBook.

The check for the boot flag should probably check for :boot; to not
just boot to keep from accidentally picking up the label.

-Jim