add edit-livecd
edit-livecd is used for opening up livecd images to add and/or edit files with out having to rebuild the iso image.

This script is useful for people who don't have the development and yum infrastructure/repos set up to build livecd's
from scratch. This is especially useful for custom livecd images.