Do lazy umounts to reduce the impact of umount issues.
By doing lazy umounts, if a umount can't be done (such as in the case of
the problem triggered by bug 617844), livecd-creator doesn't clean up after
itself. This ties up storage and loop devices.

This will mitigate problems where mounts are pinned by livecd-creator,
umounts are incorrectly done in the wrong order, or if the user is poking
around in the mount points while livecd-creator is running. It will not
fix a case where a umount is missed.

This should mitigate the symptoms of bug 509427, but not the root cause
(which is likely the problem reported in bug 617844.