Fix issue with improper exception args used by urlgrabber and pykickstart.
urlgrabber and pykickstart don't always supply both and error number and
an error message when raising an IOError exception. This was causing
an exception while processing the exception for some errors.

To prevent exceptions from insufficient arguments, the exception calls
were changed to use a single variable instead of a tuple and to use
strerror instead of the second argument.

As a temporary work around for urlgrabber and pykickstart issues, if
strerror is null, args[0] is used as a fallback so that we get a
reasonable error message.

This resolves bug 551932.