Change live.py to use the new location of the vesa splash image.
fedora-logos recently changed the location it uses for the vesa splash
image. This breaks (bug 615443) live images built using fedora-logos.
This patch changes the location that the image is fetched from to match
the one used by fedora-logos.

Note that generic-logos hasn't been changed to use the new location yet.
To accommodate this as well as building older repos on F14, a fallback
to using the old path is provided if the splash image is not found at
the new location.