Change releasever to a command line option (#667474)
The release version of the running system may not be the release you
want to build the livecd for, or the rpmdb may not be accessable from
the build environment (koji and chroot) so set the value to use for
$releasever in kickstart repo url's by passing --releasever=VER

If $releasever is used and no --releasever is passed it will try to
use the system's version. Under some circumstances (chroot environments
like koji) this isn't possible so an error will be raised.