liveimage-mount installed LiveOS with overlay
Enable standalone live-mounting of an installed LiveOS image with access to its
overlay.

Detect the presence of an overlay on a device with an installed LiveOS, and use
loop mounts and the device-mapper service to create a 'live-mounted' block
device for the attached LiveOS-bearing partition.  Provide an option to
--persist the mounting for chaining to another script.

The script continues to be a standalone tool with limited error checking. The
script requires 3 to 4 free loop devices, so is not yet suitable for use within
a booted LiveOS image.