Add title and product args (#669120)
Pass --title to set the syslinux.cfg title string
Pass --product to set the syslinux.cfg string used as part of each
kernel boot stanza.

If they are not passed it will try to split the fslabel on - and use
the first 2 elements (eg. Livecd Fedora). If that fails it will use
Linux for both.