edit-livecd: -k --kickstart option
Adds kickstart option for using kickstart file as an recipe for editing
a livecd image.

Following directives are honored:
part / --size <new rootfs size to be resized to>
bootloader --append "!opt-to-remove opt-to-add"
repo --baseurl=file://path-to-RPMs
%pre
%post
%packages

Note that repo only supports a local directory of RPMs, not a yum repo.