livecd-iso-to-disk: create partition for iso
In Fedora 17 the installer runs from the squashfs.img on the install
media. This means that the installer environment doesn't have direct
access to the install media, and it cannot mount it.

In order to support writing DVD's to USB sticks we now need to create a
second partition just for the .iso image and direct the installer to it
using the repo=UUID=... command

This patch adds creating the 2nd partition if needed and also cleans up
some of the bootloader config editing code.