remove kernel and initrd from EFI/BOOT (#811438)
The kernel and initrd don't need to be in the EFI/BOOT directory, they
are already in the isolinux dir. A side-effect of this is that the
efiboot.img and macboot.img were including them as well, causing the
image size to be larger than it needed to be.

livecd-iso-to-disk has also been adjusted to handle the new path in
the EFI boot config.