check for LIVE-REPO partition when writing DVD (#813905)
When writing a DVD to a device using livecd-iso-to-disk it needs a
second partition for the iso with the packages. Using --format sets this
up, but it was not working correctly without adding --format. It would
instead write the iso to the / of the host system.

This fixes the problem by looking for the LIVE-REPO partition as p2 of
the device and exiting if it is not found.

This also adds --inplace to sync so that devices with limited space can
be updated without reformatting.