allow for use of yum plugins during livecd creation
https://bugzilla.redhat.com/show_bug.cgi?id=813429

This adds an option ("-p", or "--plugins") to livecd-creator to allow for
the use of yum plugins during image creation. The initial usage case is
for localized live images that want to use yum-langpacks when creating their
live image so that they include the proper language packs for their locale.

Aside from the noise of propagating the commandline options down to the yum
instance, there is a bugfix where we re-call repository setup after we've
actually added the repositories; this is required so that plugins operate
properly.

Bill