copy repo data to USB for F17 (#806166)
Fedora 17 runs from the media so repo= cannot be used to mount the same
device for access to the repo so we now copy the repodata from the iso
over to the / of the USB. This requires anaconda-17.23-1 or later.

For releases before F17 the iso is still copied and repo= is setup to
point to it.

This reverts the use of 2 partitions that was used to work around this.