Write vconsole.conf directly
System Config Kickstart has been updated to use localed. This causes
problems when livecd-creator is used inside a mock (as it is with koji)
so we need a simpler way to setup the keyboard.

This writes the keyboard value directly to vconsole.conf without any
attempt to parse or verify it. If it is not included it defaults to us.

It also sets the font to latarcyrheb-sun16 which is the same thing that
Anaconda does.