litd: Add kickstart option
Add --ks kickstart option to copy a kickstart to the USB and setup
inst.ks to point to it. This can be used with a netboot or dvd iso to
make a USB stick that will start an install on boot.

Note that with the F19 release the install USB is not filtered out of
the drive list so clearpart --drives=sda needs to be used in the ks.