Check if FirewallD is installed before running ks firewall commands
It is possible for a live image to be built without FirewallD installed
in the target, either for the purpose of not configuring a firewall or
for the purpose of using an alternative.

So, now we check for if 'firewall-offline-cmd' is installed. If it
is not, then we throw a warning and ignore the configuration commands.