livecd-iso-to-disk: Adjust syslinux default menu style, as needed.
Configure the syslinux menu style to that provided in the source.
To assure a match between versions of the syslinux boot loader,
which is installed by the host, and the menu program selected by
the source, the menu program version on the host is copied to the
target device based on the menu style provided in the source.
(This means that for multi boot installations, the most recent host
and source may alter the version and menu style for all boots.)

Also, simplify selection of configuration file name.
Use more precise sed statements.
Allow better distinction of grub entries for replicated images on
  the same device.
Provide a menu option 'Return to Multi Live Image Boot Menu' on
alternate individual boot menus of a multi boot installation.