livecd-iso-to-disk+pod: Allow multi installs to live booted devices
Flash storage devices may have capacity for multiple live images.
Allow a live booted or sourced device to be both source and the
target for a live image installation. Provide an easy flag, 'live',
to request a live booted source or target.

Also, avoid an unneeded error trap in cleanup().
Preserve the device label before formatting.
Use a direct test for the presence of an md5sum in the source.
Provide a case sensitive test for EFI directories on vfat systems.
Preserve the EFI boot folder case on subsequent installations to a
  device.
Backup the previous configuration file of a multi boot installation
  that is subsequently installed without the --multi option.
Add the ro option to the mount of the source image loop device.
Use option flag labels in code; this helps reading an xtrace.
Check for sufficient target space before clearing an old image.
Account for overlaysize on --resetoverlay.
Sync only the target device filesystem.
Use more precise sed statements.
Allow better distinction of grub entries for replicated images on
  the same device.
Assure that the host Syslinux components are always written to the
  boot path, even on multi boot installations, on the logic that
  a more recent installation is likely to have newer components.
Silence the extlinux bootloader command.
Update documentation.