imgcreate: Copy gcdia32.efi in __copy_efi_files if it exists (#95)
imgcreate/live.py (__copy_efi_files): Add a required flag to the files.
Set it to True for the existing 3 files. But if set to False, just skip
the file if it is missing, and neither complain nor fail. Add
/boot/efi/EFI/*/gcdia32.efi to copy to /EFI/BOOT/grubia32.efi with
required=False.

Adds support for https://fedoraproject.org/wiki/Changes/32BitUefiSupport

Verified to do the right thing both with a Kannolo 27 x86_64 kickstart
(32-bit UEFI supported, gcdia32.efi gets copied) and with a Kannolo 26
x86_64 kickstart (32-bit UEFI not supported, gcdia32.efi is absent and
is skipped without producing an error).

Fixes #95.