Handle dnf config option showing as tuple, not list, in DNF 3.6
See https://bugzilla.redhat.com/show_bug.cgi?id=1595917 again.
This is a follow-up to 2cbd239 - after I reported the bug,
upstream changed in DNF 3.6 to present these values as tuples
rather than lists, to indicate that they cannot be mutated. So
we can't add a list to the retrieved value, like we did before,
but using += (tuple) like this seems to work both in DNF 3.6
and in older DNFs.

Signed-off-by: Adam Williamson <awilliam@redhat.com>