imgcreate/live: Check for dracut being installed to determine usability
Unfortunately, some Linux distributions patch dracut to use different
file names for the initramfs files it makes, so using the file name is
a bad indicator. Instead, we'll check to see if dracut is present on
the target system. If it is, we'll assume it's usable and go that route.