Separate the errors for no ks specified and file not found
It's more helpful to the user if we don't combine "you didn't give me a
kickstart file" and "you gave me a kickstart file, but it doesn't
exist".

In addition to giving separate error messages for the two cases, the
"you gave me a kickstart file, but it doesn't exist" says what kickstart
file it thinks it's looking for, which will help users catch syntax
errors.

Fixes #174