prevent urlgrabber.urlgrab() from overwriting ks files in current directory
When kickstart-file (--config parameter to livecd-creator) is located in
current directory, urlgrabber.urlgrab() (without filename= parameter)
will try to download the file using the basename of the path argument,
resulting in overwriting the kickstart file. The result is a kickstart file of 0 bytes.

This patch adds a filename parameter to a hidden file in the current
directory to avoid overwriting the original kickstart file. After being
parsed, the temporary file is deleted.