imgcreate/creator.py: support inserting GPG keys
Add a repo_keys argument to install(). When present, all packages
installed via %install rules must be signed with one of the
given GPG keys.

The GPG keys may become part of the final image, but the rpm macro
which enforces signature-checks does not. The macro may be added
by the image author if desired.