livecd-creator: configure dnf via --repo file or directory
The kickstart `repo` command [1] only supports a subset of the
configuration parameters available to dnf [2]. Most importantly,
kickstart does not support the `gpgkey` option. Without this
option, it is not possible to configure trusted GPG keys for
package signing. For this reason, previous versions of
livecd-creator could not check package signatures.

We can bypass this kickstart limitation by making dnf load repo
configuration from a `.repo` file or a directory of the same. This
is exactly how it works on normal, live systems.

Add a `--repo=PATH` option to livecd-creator. When provided, the
RPM repositories are sourced from this file or directory. The
repositories defined in the kickstart, if any, are ignored.

dnf's internal key-fetching logic requires that `gpgkey` be an
absolute URL. This is inconvenient for livecd-creator because the
kickstart files, repo configuration, and keys might all be stored
in some arbitrary place. Unlike stock dnf, we interpret tilde
`~/paths` and relative `./path/fragments` relative to `--repo`.

The GPG signature-checking logic is borrowed from [3].

References

1. https://docs.fedoraproject.org/en-US/fedora/latest/install-guide/appendixes/Kickstart_Sy
ntax_Reference/#sect-kickstart-commands-repo

2. https://dnf.readthedocs.io/en/latest/conf_ref.html#repo-options

3. https://github.com/rpm-software-management/dnf/blob/bc70d93609a71978e0064eb90384041da38d
c8a3/dnf/cli/cli.py#L268