imgcreate/creator.py: fix SELinux unmount order
In 67081a3657, bind mounts are unmounted before the filesystem is
relabeled with `setfiles`. The function which performs the
unmount executes

    umount /sys/fs/selinux

without first executing

    umount /sys/fs/selinux/load

This can—but may not always—lead to a "leaked a reference to the
filesystem" build error.

Fix by ensuring /sys/fs/selinux/load is unmounted first, just
like `ImageCreator.unmount()` does. We do not re-mount it
afterwards because we shouldn't need to do anything more with
SELinux after the relabel completes.