livecd-creator: add --pkgverify-level option
When set, `--pkgverify-level=LEVEL` controls the RPM macro
`%_pkgverify_level`. This macro acts as a global cryptographic
policy for `rpm`, independent of dnf.

  * a LEVEL of `all` enforces the use of trusted GPG signatures
    and strong hashes for all packages installed.

  * a LEVEL of `digest` will check only package digests.

This option acts primarily as a check on the DNF configuration
provided via `--repo`. A value of `all` ensures that all
repositories have `gpgcheck=1`, providing the user with
confidence that untrusted packages cannot be installed.

Most values for this option require --repo since it is our only
available avenue for configuring dnf GPG keys.

We must wait until after the initial GPG signature check to set
this macro. Otherwise, the installation will fail before the keys
can be imported.