Refactor to permit base-on of flattened SquashFS only images.
Use unsquashfs to decompress flattened image directly to the
  _instroot.
Avoid kickstart installations when a change root shell is requested
  in a livecd-creator session that employs a base-on image.  This
  makes sequential editing of an image quicker.

Also:
Include F2FS drivers in the initrd.
Default to zstd compression to favor more rapid decompression.
Include the dracut modules dmsquash-live-autooverlay & overlayfs
  among the defaults.
Revert the explicit omission of plymouth.
Rename dracut_args to dracut_conf_args to match their required
  syntax.
Print a couple of console messages on the activity of the script at
  longer operations.
Assure that mounts are unmounted in the necessary order.
Update documentation.